requirements.txt
pyrebase
Pillow >= 2.0.0
